refactor: rename domain name
